More filtering service refactoring
