+++ CSETWebNg/src/app/assessment/prepare/irp-summary/irp-summary.component.ts
@@ -27,7 +27,7 @@ import { AcetDashboard } from '../../../models/acet-dashboard.model';
 import { AssessmentService } from '../../../services/assessment.service';
 import { NavigationService } from '../../../services/navigation.service';
 import { ACETService } from '../../../services/acet.service';
-import { MaturityFiltersService } from '../../../services/MaturityFiltering/maturity-filters.service';
+import { AcetFilteringService } from '../../../services/filtering/maturity-filtering/acet-filtering.service';
 
 @Component({
     selector: 'app-irp-summary',
@@ -47,7 +47,7 @@ export class IrpSummaryComponent implements OnInit {
         public assessSvc: AssessmentService,
         public navSvc: NavigationService,
         public acetSvc: ACETService,
-        public acetFiltersSvc: MaturityFiltersService
+        public acetFilteringSvc: AcetFilteringService
     ) { }
 
     /**
@@ -81,7 +81,7 @@ export class IrpSummaryComponent implements OnInit {
      * 
      */
     changeInfoIrp() {
-        this.acetFiltersSvc.resetDomainFilters(this.acetDashboard.Override);
+        this.acetFilteringSvc.resetDomainFilters(this.acetDashboard.Override);
         this.changeInfo();
     }
