+++ CSETWebNg/src/app/assessment/questions/domain-block/domain-block.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { Domain } from '../../../models/questions.model';
-import { MaturityFiltersService } from '../../../services/MaturityFiltering/maturity-filters.service';
 import { QuestionsService } from '../../../services/questions.service';
 
 @Component({
