+++ CSETWebNg/src/app/assessment/questions/domain-maturity-filter/domain-maturity-filter.component.html
@@ -26,7 +26,7 @@
     [class.font-weight-bold]="l.Applicable"
     [class.font-italic]="l.Applicable"
     [class.active]="isFilterActive(l)"
-    [class.default-mat-filter]="isDefaultMatLevel(l.Level)" ngbButtonLabel>
+    [class.default-mat-filter]="acetFilteringSvc.isDefaultMatLevel(l.Level)" ngbButtonLabel>
     <input class="form-check-input" type="checkbox" ngbButton 
       [ngModel]="getNgModel(l)"
       (ngModelChange)="filterChanged(l.Level, $event)">
