+++ CSETWebNg/src/app/assessment/questions/maturity-questions/maturity-questions-acet.component.html
@@ -62,7 +62,7 @@
 
         <!-- top level (Domains) -->
         <div *ngFor="let g of groupings" class="w-100 mb-5">
-            <app-maturity-filter [domain]="g" [maturityLevels]="maturityLevels"></app-maturity-filter>
+            <app-maturity-filter [domain]="g" [maturityLevels]="maturityLevels" (change)="refreshQuestionVisibility()"></app-maturity-filter>
             <app-grouping-block [grouping]="g"></app-grouping-block>
         </div>
