+++ CSETWebNg/src/app/assessment/questions/question-block/question-block.component.ts
@@ -28,7 +28,7 @@ import { MatDialogRef, MatDialog } from '@angular/material/dialog';
 import { InlineParameterComponent } from '../../../dialogs/inline-parameter/inline-parameter.component';
 import { ConfigService } from '../../../services/config.service';
 import { AssessmentService } from '../../../services/assessment.service';
-import { QuestionFilterService } from '../../../services/question-filter.service';
+import { QuestionFilterService } from '../../../services/filtering/question-filter.service';
 
 /**
  * Represents the display container of a single subcategory with its member questions.
